feature/hilt-improvement Refactor: Move modules to the same folder
